Fix version numbers for DKMS build for ZFS
The previous version stores the ZFS code in /usr/src/zfs-0.6.5.7.r2178_4.5.4_1 causing the includes for SPL to fail during compilation of the module, since it is based on the directory name. That in turn causes mkinitcpio to fail and could potentially render the system unable to boot.

This corrects the issue while maintaining the naming of package and previous separation of _pkgver and _ver.

Former-commit-id: b5aeeb2f57827aec7cebcbb9d6b1bde8cc56cdb9